+++ ui/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@emotion/react": "^11.13.3",
     "@emotion/styled": "^11.13.0",
+    "@hookform/resolvers": "^3.9.1",
     "@mui/material": "^6.1.6",
     "@mui/system": "^6.1.6",
     "axios": "^1.7.7",
